+++ backend/turip-app/src/main/java/turip/auth/controller/dto/response/SocialLoginResponse.java
@@ -1,6 +1,5 @@
 package turip.auth.controller.dto.response;
 
-import turip.auth.service.dto.TokenResult;
 
 public record SocialLoginResponse(
         String accessToken,
